Fix bug with stackable items
Fixed a bug that was causing the script to not function correctly when
attempting to cast Alchemy on stackable items. To fix this bug, I've
added a function to determine if a item is noteable by checking the
stack size of said object. It's not a perfect solution by any means, but
it should work.